+++ oshdb-api/src/main/java/org/heigit/ohsome/oshdb/api/mapreducer/MapAggregator.java
@@ -20,7 +20,6 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.function.DoubleUnaryOperator;
 import java.util.function.Function;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.heigit.ohsome.oshdb.OSHDBBoundingBox;
@@ -30,9 +29,7 @@ import org.heigit.ohsome.oshdb.api.generic.OSHDBCombinedIndex;
 import org.heigit.ohsome.oshdb.api.generic.WeightedValue;
 import org.heigit.ohsome.oshdb.api.mapreducer.MapReducer.Grouping;
 import org.heigit.ohsome.oshdb.filter.FilterExpression;
-import org.heigit.ohsome.oshdb.osm.OSMType;
 import org.heigit.ohsome.oshdb.util.exceptions.OSHDBInvalidTimestampException;
-import org.heigit.ohsome.oshdb.util.function.OSMEntityFilter;
 import org.heigit.ohsome.oshdb.util.function.SerializableBiConsumer;
 import org.heigit.ohsome.oshdb.util.function.SerializableBiFunction;
 import org.heigit.ohsome.oshdb.util.function.SerializableBinaryOperator;
@@ -42,7 +39,6 @@ import org.heigit.ohsome.oshdb.util.function.SerializableSupplier;
 import org.heigit.ohsome.oshdb.util.mappable.OSHDBMapReducible;
 import org.heigit.ohsome.oshdb.util.mappable.OSMContribution;
 import org.heigit.ohsome.oshdb.util.mappable.OSMEntitySnapshot;
-import org.heigit.ohsome.oshdb.util.tagtranslator.OSMTagInterface;
 import org.jetbrains.annotations.Contract;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.Polygonal;
@@ -81,11 +77,11 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
    */
   MapAggregator(
       MapReducer<X> mapReducer,
-      SerializableFunction<X, U> indexer,
+      SerializableBiFunction<X, Object, U> indexer,
       Collection<U> zerofill
   ) {
-    this.mapReducer = mapReducer.map(data -> new IndexValuePair<>(
-        indexer.apply(data),
+    this.mapReducer = mapReducer.map((data, root) -> new IndexValuePair<>(
+        indexer.apply(data, root),
         data
     ));
     this.zerofill = new ArrayList<>(1);
@@ -137,9 +133,21 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
   public <V extends Comparable<V> & Serializable> MapAggregator<OSHDBCombinedIndex<U, V>, X>
       aggregateBy(SerializableFunction<X, V> indexer, Collection<V> zerofill) {
     MapAggregator<OSHDBCombinedIndex<U, V>, X> res = this
-        .mapIndex((existingIndex, data) -> new OSHDBCombinedIndex<>(
-            existingIndex,
-            indexer.apply(data)
+        .mapIndex((indexData, ignored) -> new OSHDBCombinedIndex<>(
+            indexData.getKey(),
+            indexer.apply(indexData.getValue())
+        ));
+    res.zerofill.add(zerofill);
+    return res;
+  }
+
+  // Some internal methods can also aggregate using the "root" object of the mapreducer's view.
+  private <V extends Comparable<V> & Serializable> MapAggregator<OSHDBCombinedIndex<U, V>, X>
+      aggregateBy(SerializableBiFunction<X, Object, V> indexer, Collection<V> zerofill) {
+    MapAggregator<OSHDBCombinedIndex<U, V>, X> res = this
+        .mapIndex((indexData, root) -> new OSHDBCombinedIndex<>(
+            indexData.getKey(),
+            indexer.apply(indexData.getValue(), root)
         ));
     res.zerofill.add(zerofill);
     return res;
@@ -149,12 +157,43 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
    * Sets up aggregation by another custom index.
    *
    * @param indexer a callback function that returns an index object for each given data.
+   * @param <V> the type of the values used to aggregate
    * @return a MapAggregatorByIndex object with the new index applied as well
    */
   @Contract(pure = true)
   public <V extends Comparable<V> & Serializable> MapAggregator<OSHDBCombinedIndex<U, V>, X>
       aggregateBy(SerializableFunction<X, V> indexer) {
     return this.aggregateBy(indexer, Collections.emptyList());
+  }
+
+  /**
+   * Sets up automatic aggregation by timestamp.
+   *
+   * <p>In the OSMEntitySnapshotView, the snapshots' timestamp will be used directly to aggregate
+   * results into. In the OSMContributionView, the timestamps of the respective data modifications
+   * will be matched to corresponding time intervals (that are defined by the `timestamps` setting
+   * here).</p>
+   *
+   * @return a MapAggregatorByTimestampAndIndex object with the equivalent state (settings,
+   *         filters, map function, etc.) of the current MapReducer object
+   */
+  @Contract(pure = true)
+  public MapAggregator<OSHDBCombinedIndex<U, OSHDBTimestamp>, X> aggregateByTimestamp() {
+    // by timestamp indexing function -> for some views we need to match the input data to the list
+    SerializableBiFunction<X, Object, OSHDBTimestamp> indexer;
+    if (this.mapReducer.isOSMContributionViewQuery()) {
+      final TreeSet<OSHDBTimestamp> timestamps = new TreeSet<>(this.mapReducer.tstamps.get());
+      indexer = (ignored, root) -> timestamps.floor(((OSMContribution) root).getTimestamp());
+    } else if (this.mapReducer.isOSMEntitySnapshotViewQuery()) {
+      indexer = (ignored, root) -> ((OSMEntitySnapshot) root).getTimestamp();
+    } else {
+      throw new UnsupportedOperationException(
+          "automatic aggregateByTimestamp() only implemented for OSMContribution and "
+              + "OSMEntitySnapshot -> try using aggregateByTimestamp(customTimestampIndex) instead"
+      );
+    }
+
+    return this.aggregateBy(indexer, this.mapReducer.getZerofillTimestamps());
   }
 
   /**
@@ -195,6 +234,9 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
    *
    * <p>Cannot be used together with the `groupByEntity()` setting enabled.</p>
    *
+   * @param geometries an associated list of polygons and identifiers
+   * @param <V> the type of the identifers used to aggregate
+   * @param <P> a polygonal geometry type
    * @return a MapAggregator object with the equivalent state (settings, filters, map function,
    *         etc.) of the current MapReducer object
    * @throws UnsupportedOperationException if this is called when the `groupByEntity()` mode has
@@ -212,28 +254,24 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
     }
 
     GeometrySplitter<V> gs = new GeometrySplitter<>(geometries);
-    if (this.mapReducer.mappers.size() > 1) {
-      // todo: fix
-      throw new UnsupportedOperationException(
-          "please call aggregateByGeometry before setting any map or flatMap functions"
-      );
+
+    MapAggregator<OSHDBCombinedIndex<U, V>, ? extends OSHDBMapReducible> ret;
+    if (mapReducer.isOSMContributionViewQuery()) {
+      ret = this.flatMap((ignored, root) ->
+              gs.splitOSMContribution((OSMContribution) root).entrySet())
+          .aggregateBy(Entry::getKey, geometries.keySet()).map(Entry::getValue);
+    } else if (mapReducer.isOSMEntitySnapshotViewQuery()) {
+      ret = this.flatMap((ignored, root) ->
+              gs.splitOSMEntitySnapshot((OSMEntitySnapshot) root).entrySet())
+          .aggregateBy(Entry::getKey, geometries.keySet()).map(Entry::getValue);
     } else {
-      MapAggregator<OSHDBCombinedIndex<U, V>, ? extends OSHDBMapReducible> ret;
-      if (mapReducer.isOSMContributionViewQuery()) {
-        ret = this.flatMap(x -> gs.splitOSMContribution((OSMContribution) x).entrySet())
-            .aggregateBy(Entry::getKey, geometries.keySet()).map(Entry::getValue);
-      } else if (mapReducer.isOSMEntitySnapshotViewQuery()) {
-        ret = this.flatMap(x -> gs.splitOSMEntitySnapshot((OSMEntitySnapshot) x).entrySet())
-            .aggregateBy(Entry::getKey, geometries.keySet()).map(Entry::getValue);
-      } else {
-        throw new UnsupportedOperationException(String.format(
-            MapReducer.UNIMPLEMENTED_DATA_VIEW, this.mapReducer.viewClass));
-      }
-      @SuppressWarnings("unchecked") // no mapper functions have been applied -> the type is still X
-      MapAggregator<OSHDBCombinedIndex<U, V>, X> result =
-          (MapAggregator<OSHDBCombinedIndex<U, V>, X>) ret;
-      return result;
+      throw new UnsupportedOperationException(String.format(
+          MapReducer.UNIMPLEMENTED_DATA_VIEW, this.mapReducer.viewClass));
     }
+    @SuppressWarnings("unchecked") // no mapper functions have been applied -> the type is still X
+    MapAggregator<OSHDBCombinedIndex<U, V>, X> result =
+        (MapAggregator<OSHDBCombinedIndex<U, V>, X>) ret;
+    return result;
   }
 
   // -----------------------------------------------------------------------------------------------
@@ -268,130 +306,6 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
   @Contract(pure = true)
   public <P extends Geometry & Polygonal> MapAggregator<U, X> areaOfInterest(P polygonFilter) {
     return this.copyTransform(this.mapReducer.areaOfInterest(polygonFilter));
-  }
-
-  /**
-   * Limits the analysis to the given osm entity types.
-   *
-   * @param typeFilter the set of osm types to filter (e.g. `EnumSet.of(OSMType.WAY)`)
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(String)} instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmType(Set<OSMType> typeFilter) {
-    return this.copyTransform(this.mapReducer.osmType(typeFilter));
-  }
-
-  /**
-   * Adds a custom arbitrary filter that gets executed for each osm entity and determines if it
-   * should be considered for this analyis or not.
-   *
-   * @param f the filter function to call for each osm entity
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(FilterExpression)} with {@link
-   *             org.heigit.ohsome.oshdb.filter.Filter#byOSMEntity(OSMEntityFilter)}
-   *             instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmEntityFilter(OSMEntityFilter f) {
-    return this.copyTransform(this.mapReducer.osmEntityFilter(f));
-  }
-
-
-  /**
-   * Adds an osm tag filter: The analysis will be restricted to osm entities that have this tag key
-   * (with an arbitrary value), or this tag key and value.
-   *
-   * @param tag the tag (key, or key and value) to filter the osm entities for
-   * @return a modified copy of this mapReducer (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(String)} instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmTag(OSMTagInterface tag) {
-    return this.copyTransform(this.mapReducer.osmTag(tag));
-  }
-
-  /**
-   * Adds an osm tag filter: The analysis will be restricted to osm entities that have this tag key
-   * (with an arbitrary value).
-   *
-   * @param key the tag key to filter the osm entities for
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(String)} instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmTag(String key) {
-    return this.copyTransform(this.mapReducer.osmTag(key));
-  }
-
-  /**
-   * Adds an osm tag filter: The analysis will be restricted to osm entities that have this tag key
-   * and value.
-   *
-   * @param key the tag key to filter the osm entities for
-   * @param value the tag value to filter the osm entities for
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(String)} instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmTag(String key, String value) {
-    return this.copyTransform(this.mapReducer.osmTag(key, value));
-  }
-
-  /**
-   * Adds an osm tag filter: The analysis will be restricted to osm entities that have this tag key
-   * and one of the
-   * given values.
-   *
-   * @param key the tag key to filter the osm entities for
-   * @param values an array of tag values to filter the osm entities for
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(String)} instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmTag(String key, Collection<String> values) {
-    return this.copyTransform(this.mapReducer.osmTag(key, values));
-  }
-
-  /**
-   * Adds an osm tag filter: The analysis will be restricted to osm entities that have a tag with
-   * the given key and whose value matches the given regular expression pattern.
-   *
-   * @param key the tag key to filter the osm entities for
-   * @param valuePattern a regular expression which the tag value of the osm entity must match
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   */
-  @Override
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmTag(String key, Pattern valuePattern) {
-    return this.copyTransform(this.mapReducer.osmTag(key, valuePattern));
-  }
-
-  /**
-   * Adds an osm tag filter: The analysis will be restricted to osm entities that have at least one
-   * of the supplied tags (key=value pairs or key=*).
-   *
-   * @param tags the tags (key/value pairs or key=*) to filter the osm entities for
-   * @return a modified copy of this object (can be used to chain multiple commands together)
-   * @deprecated use oshdb-filter {@link #filter(String)} instead
-   */
-  @Override
-  @Deprecated
-  @Contract(pure = true)
-  public MapAggregator<U, X> osmTag(Collection<? extends OSMTagInterface> tags) {
-    return this.copyTransform(this.mapReducer.osmTag(tags));
   }
 
   // -----------------------------------------------------------------------------------------------
@@ -828,6 +742,22 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
     }));
   }
 
+
+  // Some internal methods can also flatMap the "root" object of the mapreducer's view.
+  private <R> MapAggregator<U, R> flatMap(
+      SerializableBiFunction<X, Object, Iterable<R>> flatMapper) {
+    return this.copyTransform(this.mapReducer.flatMap((inData, root) -> {
+      List<IndexValuePair<U, R>> outData = new LinkedList<>();
+      flatMapper.apply(inData.getValue(), root).forEach(flatMappedData ->
+          outData.add(new IndexValuePair<>(
+              inData.getKey(),
+              flatMappedData
+          ))
+      );
+      return outData;
+    }));
+  }
+
   /**
    * Adds a custom arbitrary filter that gets executed in the current transformation chain.
    *
@@ -1024,12 +954,13 @@ public class MapAggregator<U extends Comparable<U> & Serializable, X> implements
   // maps from one index type to a different one
   @Contract(pure = true)
   private <V extends Comparable<V> & Serializable> MapAggregator<V, X> mapIndex(
-      SerializableBiFunction<U, X, V> keyMapper) {
-    return this.copyTransformKey(this.mapReducer.map(inData -> new IndexValuePair<>(
-        keyMapper.apply(inData.getKey(), inData.getValue()),
+      SerializableBiFunction<IndexValuePair<U, X>, Object, V> keyMapper) {
+    return this.copyTransformKey(this.mapReducer.map((inData, root) -> new IndexValuePair<>(
+        keyMapper.apply(inData, root),
         inData.getValue()
     )));
   }
+
 
   // calculate complete set of indices to use for zerofilling
   @SuppressWarnings("rawtypes")