+++ oshdb-api/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestMapReduce.java
@@ -3,6 +3,7 @@ package org.heigit.ohsome.oshdb.api.tests;
 import static org.heigit.ohsome.oshdb.OSHDBBoundingBox.bboxWgs84Coordinates;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -12,7 +13,6 @@ import org.heigit.ohsome.oshdb.api.db.OSHDBJdbc;
 import org.heigit.ohsome.oshdb.api.mapreducer.MapReducer;
 import org.heigit.ohsome.oshdb.api.mapreducer.OSMContributionView;
 import org.heigit.ohsome.oshdb.api.mapreducer.OSMEntitySnapshotView;
-import org.heigit.ohsome.oshdb.osm.OSMType;
 import org.heigit.ohsome.oshdb.util.exceptions.OSHDBTimeoutException;
 import org.heigit.ohsome.oshdb.util.function.SerializableFunction;
 import org.heigit.ohsome.oshdb.util.mappable.OSMContribution;
@@ -42,7 +42,9 @@ abstract class TestMapReduce {
     if (this.keytables != null) {
       mapRed = mapRed.keytables(this.keytables);
     }
-    return mapRed.osmType(OSMType.NODE).osmTag("highway").areaOfInterest(bbox);
+    return mapRed
+        .areaOfInterest(bbox)
+        .filter("type:node and highway=*");
   }
 
   protected MapReducer<OSMEntitySnapshot> createMapReducerOSMEntitySnapshot() throws Exception {
@@ -50,7 +52,9 @@ abstract class TestMapReduce {
     if (this.keytables != null) {
       mapRed = mapRed.keytables(this.keytables);
     }
-    return mapRed.osmType(OSMType.NODE).osmTag("highway").areaOfInterest(bbox);
+    return mapRed
+        .areaOfInterest(bbox)
+        .filter("type:node and highway=*");
   }
 
   @Test
@@ -58,7 +62,7 @@ abstract class TestMapReduce {
     // simple query
     Set<Integer> result = createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(OSMContribution::getContributorUserId)
         .uniq();
 
@@ -69,7 +73,7 @@ abstract class TestMapReduce {
     // "flatMap"
     result = createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(OSMContribution::getContributorUserId)
         .filter(uid -> uid > 0)
         .uniq();
@@ -81,7 +85,7 @@ abstract class TestMapReduce {
     // "groupByEntity"
     assertEquals(7, createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .groupByEntity()
         .map(List::size)
         .sum()
@@ -93,7 +97,7 @@ abstract class TestMapReduce {
     // simple query
     Set<Integer> result = createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps6)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(snapshot -> snapshot.getEntity().getUserId())
         .uniq();
 
@@ -102,7 +106,7 @@ abstract class TestMapReduce {
     // "flatMap"
     result = createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps6)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(snapshot -> snapshot.getEntity().getUserId())
         .filter(uid -> uid > 0)
         .uniq();
@@ -112,7 +116,7 @@ abstract class TestMapReduce {
     // "groupByEntity"
     assertEquals(5, createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps6)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .groupByEntity()
         .map(List::size)
         .sum()
@@ -124,7 +128,7 @@ abstract class TestMapReduce {
     // simple query
     Set<Integer> result = createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(OSMContribution::getContributorUserId)
         .stream()
         .collect(Collectors.toSet());
@@ -136,7 +140,7 @@ abstract class TestMapReduce {
     // "flatMap"
     result = createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(OSMContribution::getContributorUserId)
         .filter(uid -> uid > 0)
         .stream()
@@ -149,7 +153,7 @@ abstract class TestMapReduce {
     // "groupByEntity"
     assertEquals(7, createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .groupByEntity()
         .map(List::size)
         .stream()
@@ -163,7 +167,7 @@ abstract class TestMapReduce {
     // simple stream query
     Set<Integer> result = createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps6)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(snapshot -> snapshot.getEntity().getUserId())
         .stream()
         .collect(Collectors.toSet());
@@ -173,7 +177,7 @@ abstract class TestMapReduce {
     // "flatMap"
     result = createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps6)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .map(snapshot -> snapshot.getEntity().getUserId())
         .filter(uid -> uid > 0)
         .stream()
@@ -184,7 +188,7 @@ abstract class TestMapReduce {
     // "groupByEntity"
     assertEquals(5, createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps6)
-        .osmEntityFilter(entity -> entity.getId() == 617308093)
+        .filter("id:617308093")
         .groupByEntity()
         .map(List::size)
         .stream()
@@ -201,13 +205,13 @@ abstract class TestMapReduce {
   @SuppressWarnings("ResultOfMethodCallIgnored") // we only test for thrown exceptions here
   protected void timeoutMapReduce() throws Exception {
     // set short timeout -> query should fail
-    oshdb.timeoutInMilliseconds(30);
+    oshdb.timeoutInMilliseconds(1);
 
     try {
       // simple query with a sleep. would take about ~500ms (1 entity for ~5 timestamp)
       createMapReducerOSMEntitySnapshot()
           .timestamps(timestamps6)
-          .osmEntityFilter(entity -> entity.getId() == 617308093)
+          .filter("id:617308093")
           .map(delay(100))
           .count();
     } finally {
@@ -224,13 +228,13 @@ abstract class TestMapReduce {
   @SuppressWarnings("ResultOfMethodCallIgnored") // we only test for thrown exceptions here
   protected void timeoutStream() throws Exception {
     // set super short timeout -> all queries should fail
-    oshdb.timeoutInMilliseconds(30);
+    oshdb.timeoutInMilliseconds(1);
 
     try {
       // simple query
       createMapReducerOSMEntitySnapshot()
           .timestamps(timestamps6)
-          .osmEntityFilter(entity -> entity.getId() == 617308093)
+          .filter("id:617308093")
           .map(snapshot -> snapshot.getEntity().getId())
           .map(delay(100))
           .stream()